Swagger without jwt token
